requirements-api.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
proxmoxer>=2.0.1
paramiko>=3.4.0
requests>=2.31.0
sqlalchemy>=2.0.29
aiosqlite>=0.20.0
greenlet>=3.0.3
httpx>=0.27.0
pydantic-settings>=2.2.1
python-multipart>=0.0.9
python-dotenv>=1.0.1
icalendar>=5.0.12
pytest>=8.1.1
pytest-asyncio>=0.23.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
